requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth
torch
transformers
datasets
beautifulsoup4
pandas
